refactor: product-management 컴포넌트에서 product-accordion 컴포넌트 분리
